Convert all line endings to unix
A mistake was made in previous commits where this project was cloned as
a submodule in to a Windows environment, and then copied in to a Ubuntu
Docker image as part of the build process for
https://github.com/js-jslog/development-env, whereupon it was falsely
concluded that the files had been converted *from* windows *to* unix
line endings. The mistake was that even though the project was
originally developed on Windows the line endings would have been
converted back to unix during commit. In the Docker copy situation they
were copied to the image in their Windows format. No commit was involved
so no conversion occurred. Then in the container context the git config
would naturally assume that no conversion was required.