Get initial node file debugging working
There is a lot included in this commit which hasn't actually been tested
or isn't actually working yet.

The only thing which has proven to be working is the pwa-node launch.
The node2 stuff doesn't appear to be working and might not even be
needed (might be deprecated in favour of pwa-node).

Additionally I've installed some ui things which haven't been tested
yet.

But at the end of the day, I just wanted to get something committed
because it's been quite hard to get to this point.