Manual install of vscode-js-debug & use nvim-dap-vscode-js
This then means that mason-nvim-dap is no longer used, but I'll leave it
there anyway.

This seems to be a more common way of doing things for node. The
nvim-dap-vscode-js package might make certain things easier. I was
reluctant to use an extra layer of abstraction, but until I get the
attach functionality working I might benefit from it.