Clean up the dap config a little bit
I'm actually too much of a coward to remove all the comments in the
dap-vscode-js configuration. I think this is because I want to be
reminded that there are options here if I need to extend the usage at
some point. I shouldn't need the comments for that, but I'm just a bit
uncertain about this plugin, and I've been unsuccessful at replacing it
with hand rolled config.