Confirm prettier setup
After a lot of experimentation I can confirm that I'm happy with this
now. There are performance benefits in moving to prettierd but I don't
have a clear need for that yet, and the downsides are the need to
specify the prettier config in an environment variable which has it's
own limitations.