Move most lazy dependencies to deps section
It's possible that the LSP and DAP sections could be restructured to a
dependency format instead of the simple list of plugins that it is at
the moment. But since all those that remain are only referenced once it
isn't that much less tidy to leave it as it is and it's safer than
refactoring those sections which are quite complicated and difficult to
check while I'm still not that familiar with them.

I've also not added versions to all the moved deps yet. That'll be
covered in an upcoming commit.