Convert from tag to commit where tags are old
The tags for plenary and treesitter are too old. In fact the tag for
treesitter is so old that the neotest-jest plugin won't find tests
when treesitter is using that commit.

I'd rather use tags than commits because if an autor uses tags it does
suggest that they are more certain about the stability at those stages.