Remove Noice for now
This doesn't seem to be working that well at the moment. The
configuration items don't seem to have an effect and I don't like how
intrusive the popup notifications are.

I might reintroduce this when I have upgraded the neovim build version
if that's the reason that it's not working well. Speculative, but
possible.